+++ R/latex-dependencies.R
@@ -16,7 +16,8 @@ huxtable_latex_dependencies <- list(
   list(name = "calc"),
   list(name = "tabularx"),
   list(name = "threeparttable"),
-  list(name = "wrapfig")
+  list(name = "wrapfig"),
+  list(name = "fontspec")
 )